code refactoring, bug fix
