+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_packages
 
-VERSION = "1.3.3"
+VERSION = "1.4.0"
 DESCRIPTION = "A concurrent python download manager"
 with open("README.md", "r") as f:
     LONG_DESCRIPTION = f.read()
